Wire up product-selection from a model into the launch configurations. Currently only enabled for the Java backend, it's grayed out for Maude.
There may be a slight hitch with how the builder/ABSNature manage the model: Model.flattenProduct() mangles the model, so the code generator should probably use a deep copy, as the nature will hold on to the model reference. But calling Model.fullCopy() will later trigger an exception in the Java code-gen. So for now it'll just call copy(), until we can see where it breaks.

git-svn-id: https://repos.hats-project.eu:444/svn/hats/Tools/ABS/trunk@11678 97f52227-7fa9-4375-a72d-c1c733b1f48f