#407 Handle AwaitAsyncCall uniformly for all backends (by hitting it with rewriting rules):
Any such call gets a corresponding temporary variable declaration with the AsyncCall and an await-statement. The actual call is then replaced with a GetExp on the temp. variable. I figured GenerateCoreAbs was the right place to put it since it's not really used, but captures the spirit.

Requires a bit of JastAdd machinery, as we need semantic info on the return type of the call for the variable declaration. I hope I got the the logic in Type right.

The MaudeBackend has native support, so switch the rules off there.

git-svn-id: https://repos.hats-project.eu:444/svn/hats/Tools/ABS/trunk@19961 97f52227-7fa9-4375-a72d-c1c733b1f48f