Maude backend: implement die statement
We implement a "half-synchronous" semantics: all processes executing on
the object get rewritten synchronously to immediately execute abort the
next time they are scheduled, but the state of their futures is not
immediately changed from Waiting.

To do: decide which error such terminated processes should emit.  At the
moment, they throw the same error as the original "die" statement - we
might want to use a generic "object crashed" error instead.

To do: do synchronous method calls have the same semantics?  If a
process running on O synchronously called a method on O' (on the same
cog) which suspends, and O gets subsequently killed before the process
on O' resumes, in our current semantics O' gets to finish running that
process (and then filling a future that will never be read).  In other
implementations, the process running in the context of O' might get
killed immediately along with the rest of O.

To do: deactivate the dead object in some way so it doesn't accept new
method calls