Maude backend: implement try-catch-finally
Errors are represented by strings; this is not their final form.

An uncaught die statement, or any error not explicitly raised via abort,
will kill the whole object and all processes.  Such errors can be caught
and re-thrown via abort to terminate the current process only.

Synchronous method calls will not have their caller scheduled
immediately; this was easiest to implement and will be revisited once we
decide on the semantics of synchronous method calls wrt exceptions.  The
semantic consequence is that a synchronous method call has to be treated
as a suspension point wrt program correctness - but we have to do that
anyway since the called method might suspend, so the invariant has to
hold at the calling point.