Fixes gc deadlock with uninitialized object tasks.
If objects were not activated, processes would block waiting for them to
start. But they were not capable of responding to garbage collector
messages. To fix this, uninitialized objects respond with an uninitialized
message, which puts the task in a blocking state where it can respond to
garbage collector messages.