Fixes deadlock stopping COGs with no runnable tasks.
When there were no runnable tasks, the next state was always a state where
nothing is running and a task would be scheduled. Thus it discarded the
new state's running field, which is set when stopping the world.

It might be better to always set the running field when receiving
messages, rather than handle the state's in an additional case.

Also changes getting references to use a union of sets gotten from tasks.
This is done by the COG to reduce the message size sent back to the
process waiting for the result.