Fixes objects being hidden from the GC.
If an object is created, and the world stops while its init block is being
run and the creating task has not resumed, the object reference was not
available to the creator in its ready loop. When the init block finishes
execution, there are no more tasks on the object, and thus neither the
creator or any mutator's elsewhere have reference to the newly created
object.