Allow non-unique accessor functions for a data type
A data type may have multiple constructors, each of which can have multiple
named (or unnamed) accessor function. We want the type checker to allow
constructors to reuse named accessor functions if the name refers to the same
type.

Example: Let the type checker accept the following expression:

      data Foo = Bar(Bool isTrue, String name) | Baz(String name);

But do not accept the following:

    data Foo = Bar(Bool isTrue, String name) | Baz(Int name);