Fix for tests, which do not terminate (#254)
* Added project `excludeTests` property to frontend build gradle to be able to exclude tests dynamically
This will help me to detect the test which does not terminate for me.

* Extended build.gradle property `excludeTests` to accept a comma separated list

* Erlang backend tests: TimeoutThread class now uses destroyForcibly instead of destroy().
It seems Deadlock tests do not terminate, since they are supposed to be killed after a timeout. However, it appears erlang sometimes needs to be killed using SIGKILL (destroyForcibly()), since it does not react to SIGTERM (destroy())
However, the Java standard does not enforce the use of SIGKILL in
destroyForcibly. Therefore this is not guaranteed to work on all
plattforms.

* Reverted the `destroyForcibly` commit and adapted input reading method instead.
It turns out, killing the process was not the issue. Instead the main
thread blocked because it still tried to read input from the killed
process. These changes should deal with this problem.

* Fix for Erlang tests not recording their result, as pointed out by @rudi

* Fix for missed compiler errors in last commit

* Missing braces on altered if statement caused input reading loop for erlang tests to abort too early

* Fixed input stream not being read until the end if tests terminate before the timeout, but still fail due to assertion etc.