Fix crash with Cost annotation on method call
The Cost annotation was used to consume resources at the call site, but
also to set the cost attribute of the task.

This was based on a misunderstanding of the cost attribute, which is an
attribute of the method itself, not the call.  The cost attribute is
also unused since we do not specify method execution costs: instead we
simulate method execution costs via duration statements and statement
costs.