Convert min, max, abs to built-in functions
The abs function returned Rat for values of type Int, and did not work
for Float values at all, due to limitations in the ABS type system.

The functions min and max needed the two arguments to be of compatible
type both in abs and the backend, which would not work when making the implementation of integers and rational numbers distinct.