Work around erlang backend compilation failure
On MacOS versions 15 and 26 with XCode 26, rebar returns a non-zero
exit code when compiling the Erlang backend during build, possibly
related to compiling the native-code parts of the esqlite library.
Ignore the build failure for now.

This fix leaves the erlang backend unable to query sqlite databases.

Other platforms are not affected (tested on Linux).

Reported in #379